parametricBO/BGS_opt/requirements.txt
aiohttp==3.6.2
antlr4-python3-runtime==4.8
argcomplete==1.12.3
argon2-cffi==21.1.0
async-timeout==3.0.1
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==4.1.0
boto==2.49.0
certifi==2020.4.5.1
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.5
click==7.1.2
cloudpickle==2.0.0
colorama==0.4.3
crcmod==1.7
cryptography==3.4.8
cycler==0.10.0
debugpy==1.4.3
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
fasteners==0.16.3
filelock==3.0.12
future==0.18.2
gcs-oauth2-boto-plugin==2.7
google==2.0.3
google-apitools==0.5.32
google-reauth==0.1.1
grpcio==1.29.0
gsutil==4.68
gym==0.20.0
higher==0.2.1
httplib2==0.19.1
hydra-core==1.1.1
idna==2.9
importlib-metadata==1.6.0
importlib-resources==5.2.2
ipdb==0.13.9
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.4
jedi==0.18.0
Jinja2==3.0.1
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
kiwisolver==1.3.2
learn2learn==0.1.6
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.6
mpmath==1.2.1
msgpack==0.6.2
multidict==4.7.6
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.18.4
oauth2client==4.1.3
omegaconf==2.1.1
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.12.2
ptyprocess==0.7.0
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.2
pytz==2021.1
pyu2f==0.1.5
PyYAML==5.3.1
pyzmq==22.2.1
qpth==0.0.15
qtconsole==5.1.1
QtPy==1.11.0
ray==0.8.5
redis==3.4.1
requests==2.23.0
retry-decorator==1.1.1
rsa==4.7.2
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.11.2
Send2Trash==1.8.0
simpy==4.0.1
six==1.15.0
soupsieve==2.0.1
sympy==1.8
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.1.0
tinydb==4.5.1
tinyrecord==0.2.0
toml==0.10.2
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
typing==3.7.4.3
typing-extensions==3.10.0.2
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
yarl==1.4.2
zipp==3.1.0
